Refactor engine structure and add config system
Major refactor of the engine structure: introduces a new configuration system (scythe.ini, config.cs), command-line interface (cli.cs), and core engine logic (core.cs). Replaces and reorganizes editor and runtime entry points, updates object/type instantiation, and improves modularity. Adds light support, enhances model/material handling, and updates transform and animation logic. Cleans up and renames files for clarity, migrates types, and updates README with new usage and configuration instructions. Removes obsolete files and restructures the project for better maintainability and extensibility.

+++ scythe/types/color.cs
@@ -21,7 +21,7 @@ public static partial class ext {
         
         return new(value.R, value.G, value.B, value.A);
     }
-    
+  
     public static System.Numerics.Vector4 to_vector4(this color color) {
         
         return new(color.r, color.g, color.b, color.a);
@@ -41,6 +41,6 @@ public static partial class ext {
 
         var bytes = color.to_bytes();
         
-        return ((uint)bytes.x << 24) | ((uint)bytes.y << 16) | ((uint)bytes.z << 8) | (uint)bytes.w;
+        return ((uint)bytes.x << 24) | ((uint)bytes.y << 16) | ((uint)bytes.z << 8) | bytes.w;
     }
 }
